Fix after @Osmose comments on PR.
